core/requirements.txt
Flask==1.0.2
flask_injector==0.11.0
pycorenlp==0.3.0
python_twitter==3.5
